requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==2.0.0
alembic==1.4.1
altair==4.1.0
aniso8601==9.0.1
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
async-timeout==3.0.1
attrs==21.2.0
azure-core==1.17.0
azure-storage-blob==12.8.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blessings==1.7
blinker==1.4
boto3==1.18.24
botocore==1.21.24
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
coverage==5.5
cryptography==3.4.7
databricks-cli==0.15.0
debugpy==1.5.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.1
docker==5.0.0
dotmap==1.3.24
elastic-apm==6.3.3
elasticsearch==7.10.0
entrypoints==0.3
faiss-cpu==1.7.1.post2
farm==0.8.0
fastapi==0.68.0
filelock==3.0.12
Flask==1.1.4
Flask-Cors==3.0.10
flask-restplus==0.13.0
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.31.2
google-auth==1.35.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
greenlet==1.1.1
grpcio==1.37.1
grpcio-tools==1.34.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
huggingface-hub==0.0.8
idna==3.2
install==1.3.4
ipykernel==6.5.1
ipython==7.30.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
langdetect==1.0.9
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.13.1
mmh3==3.0.0
more-itertools==8.8.0
msgpack==1.0.2
msrest==0.6.21
multidict==5.1.0
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
nltk==3.6.2
notebook==6.4.6
numpy==1.21.2
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.2.0
pluggy==0.13.1
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
prompt-toolkit==3.0.23
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
py-spy==0.3.8
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.10.0
pymilvus==1.1.2
Pympler==0.9
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-docx==0.8.11
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==22.3.0
querystring-parser==1.2.4
ray==1.5.0
rdflib==6.0.0
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
sklearn==0.0
smmap==4.0.0
SPARQLWrapper==1.8.5
SQLAlchemy==1.4.23
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.1
starlette==0.14.2
streamlit==1.2.0
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tika==1.24
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.8.1
torch-scatter==2.0.9
tornado==6.1
tox==3.24.2
tqdm==4.62.1
traitlets==5.1.1
transformers==4.6.1
typing-extensions==3.10.0.0
tzdata==2021.5
tzlocal==4.1
ujson==4.1.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.14.0
validators==0.18.2
virtualenv==20.7.2
watchdog==2.1.6
wcwidth==0.2.5
weaviate-client==2.5.0
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==0.16.1
widgetsnbextension==3.5.2
yarl==1.6.3
